添加任务历史处理人

+++ assess-app/src/main/java/com/copower/pmcc/assess/controller/project/ProjectInfoController.java
@@ -376,7 +376,7 @@ public class ProjectInfoController {
     @RequestMapping(value = "/getTotalPlanDetails", name = "计划总数", method = RequestMethod.POST)
     public HttpResult getTotalPlanDetails(Integer planId) {
         try {
-            Integer total = projectPlanDetailsService.getTotalPlanDetails(planId);
+            Long total = projectPlanDetailsService.getTotalPlanDetails(planId);
             return HttpResult.newCorrectResult(total);
         } catch (Exception e) {
             baseService.writeExceptionInfo(e, "取得计划总数异常");
